Use absolute URLs to public service on landing page
For alternate deployments that do not use the default data-models repo,
the example links to the OMOP data model are broken. Using URLs to the
public service ensures these will always work.

A more desirable approach may be to default to a local data model if at
least one is indexed by the service.

Signed-off-by: Byron Ruth <b@devel.io>